Adding passengers to flight, added JavaDocs

+++ src/main/java/com/mslup/lot/lotcrud/filter/FlightFilterCriteria.java
@@ -1,15 +1,40 @@
 package com.mslup.lot.lotcrud.filter;
 
-import jakarta.persistence.criteria.Predicate;
 import java.time.OffsetDateTime;
 import lombok.Builder;
 
+/**
+ * Klasa reprezentująca kryteria filtrowania lotów.
+ */
 @Builder
 public class FlightFilterCriteria {
+    /**
+     * Kod lotniska początkowego.
+     */
     public String originAirport;
+
+    /**
+     * Kod lotniska docelowego.
+     */
     public String destinationAirport;
+
+    /**
+     * Data początkowa.
+     */
     public OffsetDateTime dateFrom;
+
+    /**
+     * Data końcowa.
+     */
     public OffsetDateTime dateTo;
-    public Long seatsCountFrom;
-    public Long seatsCountTo;
+
+    /**
+     * Minimalna liczba miejsc.
+     */
+    public Integer seatsCountFrom;
+
+    /**
+     * Maksymalna liczba miejsc.
+     */
+    public Integer seatsCountTo;
 }
